refs #51 The design of the task of projects has been reworked. The ProjectTasks are not Activities anymore ; but instead they are linked to Activities which replace the Workingperiods ; the ProjectTasks still have start/end/duration/description fields.
The Contact linked to a Resource can now be changed ; but a Resource cannot be deleted if it is related to Activity (before, the related WorkingPeriods were deleted too).

API breaking:
 - The model ProjectTask does not inherit from Activity anymore (it inherits CremeEntity directly).
   the fields 'title', 'start', 'end' & 'duration' have been copied from the Activity part.
 - The block TaskWorkingPeriodsBlock has been replaced by TaskActivitiesBlock.
   The template file 'block_working_periods.html' has been renamed to 'block_activities.html'.
 - The model WorkingPeriod, the form WorkingPeriodForm & the views form views/workingperiod.py have been removed.
   The method ProjectTask.get_working_periods() has been removed too ; use the property 'related_activities' instead.
 - Resource
   - hourly_cost is not nullable anymore, with 0 as default.
   - task/linked_contact fields are not editable anymore.
 - ProjectTask views:
   - The argument of detailview() has been renamed to 'task_id' (from 'object_id').
   - delete() has been removed (the generic one is used).

+++ creme/projects/views/resource.py
@@ -2,7 +2,7 @@
 
 ################################################################################
 #    Creme is a free/open-source Customer Relationship Management software
-#    Copyright (C) 2009-2014  Hybird
+#    Copyright (C) 2009-2015  Hybird
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as published by
@@ -23,11 +23,14 @@ from django.shortcuts import get_object_or_404
 from django.http import HttpResponse
 
 from creme.creme_core.auth.decorators import login_required, permission_required
+from creme.creme_core.core.exceptions import ConflictError
+from creme.creme_core.models import Relation
 from creme.creme_core.utils import get_from_POST_or_404
 
+from ..constants import REL_SUB_PART_AS_RESOURCE
 from ..forms.resource import ResourceCreateForm, ResourceEditForm
 from ..models import Resource
-from .utils import _add_generic, _edit_generic
+from .utils import _add_generic, _edit_generic #TODO: these generic views are only here => 'inline' them
 
 
 @login_required
@@ -48,6 +51,13 @@ def delete(request): #TODO: generic delete ??
 
     request.user.has_perm_to_change_or_die(resource.task)
     #request.user.has_perm_to_delete_or_die(resource) #beware to change template if uncommented
+
+    if Relation.objects.filter(subject_entity=resource.linked_contact_id,
+                               type=REL_SUB_PART_AS_RESOURCE,
+                               object_entity__in=[a.id for a in resource.task.related_activities],
+                              ) \
+                       .exists():
+        raise ConflictError(_('This resource cannot be deleted, because it is linked to activities.'))
 
     resource.delete()
